backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
chromadb
google.genai
python-dotenv
pydantic